Move match guard helper to bound node facts
